修复用户端隧道流量配额显示为 0 GB 的问题
UserTunnelDetailDto.flow 定义为 Long，导致数据库中的
小数值（如 0.5 GB）被截断为 0，用户看到流量配额显示 0 GB。
将类型改为 Double 与 UserTunnel 实体保持一致。

+++ springboot-backend/src/main/java/com/admin/common/dto/UserPackageDto.java
@@ -60,7 +60,7 @@ public class UserPackageDto {
         private Integer tunnelId;
         private String tunnelName;
         private Integer tunnelFlow;  // 隧道流量计算类型（1-单向，2-双向）
-        private Long flow;           // 隧道流量配额(GB)
+        private Double flow;         // 隧道流量配额(GB)
         private Long inFlow;         // 隧道已用入站流量(字节)
         private Long outFlow;        // 隧道已用出站流量(字节)
         private Integer num;         // 隧道转发数量配额
